# specs/v1.openapi.yaml
openapi: 3.0.3
info:
  title: Fox Converter
  description: Fox Converter API for data conversion
  version: 1.0.0

servers:
  - url: 'https'

paths:

  /v1/sellers/id/{id}:
    get:
      description: Get seller data from DB by id
      operationId: getSeller
      security:
        - Keycloak:
          - sellers:read
      parameters:
        - name: id
          required: true
          in: path
          description: Id of seller
          schema:
            type: string
            format: URLs
          x-go-param-name: URLParams
      responses:
        '200':
          description: Seller Data in JSON Format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Seller"

  /v1/sellers/all:
    get:
      description: Get all seller's Ids
      operationId: GetSellersAll
      security:
        - Keycloak:
          - sellers:read
      responses:
        '200':
          description: Seller Data in JSON Format
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Seller"

  /v1/sellers/search:
    get:
      description: Get seller data from DB by name
      operationId: getSellersByName
      security:
        - Keycloak:
          - sellers:read
      parameters:
        - name: name
          required: true
          in: query
          description: name of seller
          schema:
            type: string
      responses:
        '200':
          description: Seller Data in JSON Format
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Seller"
    post:
      description: Create new Seller
      operationId: postSeller
      security:
        - Keycloak:
          - sellers:write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Seller'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Seller"


components:
  securitySchemes:
    Keycloak:
      type: oauth2
      flows: 
        authorizationCode:
          authorizationUrl: 'http://'
          tokenUrl: 'http://'
          scopes:
            sellers:read: "Read sellers"
            sellers:write: "Change and delete sellers"
      description: OAuth2 client credentials grant flow with Keycloak
  
  schemas:
    
    Seller:
      type: object
      description: Seller Master Data
      properties:
        id:
          type: string
          format: uuid
        contact:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        ogrn:
          type: string
        inn:
          type: string
        city:
          type: string
        rating:
          type: number
          format: float
        ratingsCount:
          type: integer
          format: int32
        imageNames:
          type: array
          items:
            type: string
        created:
          type: string
        updated:
          type: string